Refactor ApiHandler to support cross-version async compatibility.
* API handler classes now in own subpackage
* Implemented BaseApiHandler class common to all versions
* Non-async ApiHandler and async AsyncApiHandler now in separate modules
* API route implementation now imports sync/async handler depending on
  python version
